backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
httpx==0.26.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
openai==1.12.0